pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "pptx-cli"
dynamic = ["version"]
description = "Template-bound PowerPoint generation for enterprise decks"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [
  { name = "Thomas Rohde" }
]
keywords = ["pptx", "powerpoint", "cli", "openxml", "enterprise"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Office/Business",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
  "typer>=0.15.1",
  "rich>=13.9.4",
  "pydantic>=2.10.6",
  "PyYAML>=6.0.2",
  "jsonschema>=4.23.0",
  "lxml>=5.3.1",
  "python-pptx>=1.0.2"
]

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.11.0",
  "pyright>=1.1.397"
]

[project.scripts]
pptx = "pptx_cli.cli:main"

[tool.hatch.version]
path = "src/pptx_cli/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/pptx_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["pptx_cli"]
